Hide internal types in image module
